% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ak_coords}
\alias{ak_coords}
\title{Alaskan Coordinates}
\format{A data frame with 2 variables: \code{lon}, \code{lat}}
\usage{
ak_coords
}
\description{
WGS84 coordinates of select communities in Alaska.
}
\keyword{datasets}
