% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorise_flower_scores.R
\name{categorise_flower_scores}
\alias{categorise_flower_scores}
\title{Classify flower colour scores.}
\usage{
categorise_flower_scores(red, yellow)
}
\arguments{
\item{red}{Vector of red scores between 0 and 5.}

\item{yellow}{Vector of yellow scores between 0 and 4}
}
\value{
Data frame listing flower colour phenotype (yellow, white, weak orange,
full orange, weak rea and full red), and genotypes for \emph{Rosea}
and /emph{Sulfurea}.
}
\description{
Classify snapdragon flower colour and genotype using red and yellow colour scores.
}
\details{
Hybrid zone \emph{Antirrhinum majus} plants segregare for the loci \emph{Rosea}
and \emph{Sulphurea} controlling anthocyanin (red or magenta) and aurone
(yellow) pigmentation. `categorise_flower_scores` classifies flower colour into
one of six phenotype classes and returns genotypes for pigmentation loci
based on the scoring scheme developed by Annabel Whibley in her thesis.
}
\references{
Whibley, Annabel (2004), Molecular and genetic variation
underlying the evolution of flower colour in Antirrhinum. Ph.D. thesis,
University of East Anglia.
}
\author{
Tom Ellis, David Field
}
