% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_frequency.R
\name{density_frequency}
\alias{density_frequency}
\title{Plant density and phenotypic frequency}
\usage{
density_frequency(focal, population, scale, focal_phenotypes = NULL,
  population_phenotypes = NULL, density_function = "gaussian")
}
\arguments{
\item{focal, population}{Positional information about a set of focal plants
and plants from the wider population of neighbours. This may be a vector of
positions, or (more likely) a data.frame of coordinates, with a column for
each axis. Arbitrary numbers of coordinate axes are allowed.}

\item{scale}{Float indicating the scale at which to look at neighbours.
If `density_function` is set to `gaussian`, this is the standard deviation of
the Gaussian function to use.
If `density_function` is set to `radius`, this is the radius within which
plants are classified as neighbours.}

\item{focal_phenotypes, population_phenotypes}{Optional vectors of phenotype
data for each individual in focal and population.}

\item{density_function}{String indicating whether to calculate density using
a Gaussian function of distance, or counting the number of neighbours within
a certain radius. Must take the values 'gaussian' or 'radius'. See
`Description` for details.}
}
\value{
A vector of densities for each plant in focal. If phenotypes are
supplied, a data.frame of densities and phenotypic frequencies are returned.
}
\description{
For a set of flocal plants you are interested in, calculate the density of
neighbouring plants, and (optionally) their local phenotypic frequency.
}
\details{
Density can be calculated two ways. Most simply, we can count the number of
neighbours within a given radius, which is simple to interpret, but gives
coarse resolution. I have found that small changes to the radius can have
enormous effects on estimates of density.

Alternatively, one can apply a Gaussian function to distances using the
standard deviation parameter as a scale. The density of plants around a
focal plant is then the sum of this function on distances to all neighbours
in the population. This has the advantage of smoothing effects with distance,
and ensuring that nearer neighbours have a greater influence than mor distant
neighbours. However, interpretation is not as straightforward as using a
fixed radius.

Phenotypic frequency is defined as the density of neighbours of the same
phenotype relative to the density of all neighbours. Thus, it is only defined
for categorical phenotypes.
}
\references{
Ellis T (2016), "*The role of pollinator-mediated selection in the
maintenance of a flower color polymorphism in an Antirrhinum majus hybrid
zone*", PhD thesis, IST Austria, available at https://repository.ist.ac.at/526/
}
\author{
Tom Ellis
}
