% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_ramet.R
\name{match_ramet}
\alias{match_ramet}
\title{Value matching for RametIDs}
\usage{
match_ramet(input, plantIDs, rametIDs)
}
\arguments{
\item{input}{Vector of AntspecIDs for plants you would like to extract for.}

\item{plantIDs}{Vector of primary genet names.}

\item{rametIDs}{Vector of RametID strings.}
}
\value{
And vector of integers giving the position of each element in input
in the vector of rametIDs. Use this to subset the Antspec data.table.
}
\description{
`match_ramet` returns a vector of the positions of (first) matches of
individual Antspec IDs in in a vector of RametIDs.
}
\details{
We often have data for a ste of plants from the field, but the Antspec IDs
might not match the unique IDs after duplicate genotypes have been filtered.
Instead, we get a table with a column "RametID" that gives names of likely
duplicates as a string, which is difficult to parse.

`match_ramet` works like the base R function `match` in that it finds the
(first) position of your AntspecIDs in the column of RametIDs.

Note that `match_ramet` assumes that there are
no more than five names in any entry in `RametIDs`.
}
\author{
Tom Ellis
}
