% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrf_get.R
\name{wrf_cells}
\alias{wrf_cells}
\title{Determine cells in WRF grid corresponding to supplied data.frame of coordinates}
\usage{
wrf_cells(coords)
}
\arguments{
\item{A}{two-variable data.frame of WGS84 coordinates (columns lon, lat)}
}
\description{
Determine cells in WRF grid corresponding to supplied data.frame of coordinates
}
