% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrf_get.R
\name{wrf_get}
\alias{wrf_get}
\title{Access the "Historical and Projected Dynamically Downscaled Climate Data
for the State of Alaska and surrounding regions" datasets produced by SNAP}
\usage{
wrf_get(nc_fns, coords)
}
\arguments{
\item{coords}{A two_variable data.frame of WGS84 coordinates}

\item{nc_fn}{A netcdf file name (path) containing the downscaled data,
or vector of such filenames}
}
\description{
Access the "Historical and Projected Dynamically Downscaled Climate Data
for the State of Alaska and surrounding regions" datasets produced by SNAP
}
\examples{
#
# coordinates for Nome, AK
#
nome <- ak_coords[1, ]
fn <- "t2min_daily_wrf_GFDL-CM3_historical_1979.nc"
wrf_get(fn, "t2min", nome)
}
