% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrf_get.R
\name{wrf_transform}
\alias{wrf_transform}
\title{Transform coordinates to WRF projection}
\usage{
wrf_transform(coords, epsg = 4326, proj4str = NULL)
}
\arguments{
\item{coords}{a data.frame of lon, lat coordinates}

\item{epsg}{An EPSG code for projection of input coordinates}

\item{proj4str}{a PROJ4 string (must be supplied if epsg isn't)}
}
\description{
Transform coordinates to WRF projection
}
