% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrf_get.R
\name{wrf_xy}
\alias{wrf_xy}
\title{return df of x, y values of centerpoints of square grid of size (n*2 + 1)^2 centered on coords}
\usage{
wrf_xy(coords, n = 2, ret_mask = FALSE)
}
\arguments{
\item{coords}{A pair of lon/lat coords (if data.frame, first row will be used)}

\item{n}{Whole number specifying size of desired grid (square with (n*2 + 1)^2 cells)}

\item{ret_mask}{Logical specifying whether to return the land mask values}
}
\description{
return df of x, y values of centerpoints of square grid of size (n*2 + 1)^2 centered on coords
}
