% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrf_get.R
\name{wrf_get}
\alias{wrf_get}
\title{Access the "Historical and Projected Dynamically Downscaled Climate Data
for the State of Alaska and surrounding regions" datasets produced by SNAP}
\usage{
wrf_get(nc_fns, coords, shift = NULL, rc = FALSE, use_par = FALSE)
}
\arguments{
\item{nc_fns}{A netcdf file name (path) containing the downscaled data,
or vector of such filenames}

\item{coords}{A two-variable data.frame of WGS84 coordinates (columns lon, lat)}

\item{shift}{A vector of horizontal (1st element) and vertical (2nd element) grid cell positions to shift before querying WRF output}

\item{rc}{A logical arg indicating whether supplied coords are row/column values}

\item{use_par}{Logical indicating parallel::mclapply should be used internally}
}
\description{
Access the "Historical and Projected Dynamically Downscaled Climate Data
for the State of Alaska and surrounding regions" datasets produced by SNAP
}
\examples{
#
# coordinates for Nome, AK
#
nome <- ak_coords[1, ]
fn <- "t2min_daily_wrf_GFDL-CM3_historical_1979.nc"
wrf_get(fn, "t2min", nome)
}
